Back-End/requirements.txt
Flask==2.1.0
flask-cors==3.0.10
mysqlclient==2.1.0
SQLAlchemy==1.4.35
google-auth-oauthlib==0.5.1
google-api-python-client==2.43.0
google-auth-httplib2==0.1
waitress==2.1.1
distro==1.7.0